Trying Int8 and action of group, not improving D:

+++ src-exe/IO_Visualizator/InputBandagedCube.hs
@@ -4,9 +4,10 @@ import Moves
 import Cube
 import Bandaged
 import qualified Data.Set as S
+import Data.Int (Int8)
 
 -- | Creates a new Bandaged Cube, with a standard cube and the restrictions
-newBandagedCube :: Cube -> [[Int]] -> BandagedCube
+newBandagedCube :: Cube -> [[Int8]] -> BandagedCube
 newBandagedCube cubeOrigin blocks = BandagedCube {stdCube = cubeOrigin, restrictions = postProcessedBlocks}
     where
         blocksSet = doubleListToDoubleSet blocks
@@ -19,12 +20,12 @@ doubleListToDoubleSet :: Ord a => [[a]] -> S.Set (S.Set a)
 doubleListToDoubleSet xs = S.fromList (map S.fromList xs)
 
 -- | Join (union) the sets that are not disjoint.
-canonicSets :: S.Set (S.Set Int) -> S.Set (S.Set Int)
+canonicSets :: S.Set (S.Set Int8) -> S.Set (S.Set Int8)
 canonicSets ss = genUnions (S.toList ss) (doubleListToDoubleSet [[]])
 --Maybe it can be improved with folds, and not convert to lists back again
 
 ---- Auxiliar of canonic set. Takes a list of sets, makes the unions recursively
-genUnions :: [S.Set Int] -> S.Set (S.Set Int) -> S.Set (S.Set Int)
+genUnions :: [S.Set Int8] -> S.Set (S.Set Int8) -> S.Set (S.Set Int8)
 genUnions [] xs = S.delete (S.empty) xs 
 genUnions (s0:rest) temp = genUnions rest (S.insert unitedSet notCombining)
     where
@@ -33,17 +34,17 @@ genUnions (s0:rest) temp = genUnions rest (S.insert unitedSet notCombining)
         --Generates the union of the non-disjoint sets
 
 --Tries to expand all the blocks
-expandBlocks :: S.Set (S.Set Int) -> S.Set (S.Set Int)
+expandBlocks :: S.Set (S.Set Int8) -> S.Set (S.Set Int8)
 expandBlocks setOfBlocks = S.map expand1Block setOfBlocks
 
 --Expand a single block, with all the possible turns
-expand1Block :: S.Set Int -> S.Set Int
+expand1Block :: S.Set Int8 -> S.Set Int8
 expand1Block block = pieceIntersections (expansions)
     where
         expansions = map (expandBlockTurn block) [R,U,L,D,B,F]
     
 --Given a block and a Face, returns the expanded block
-expandBlockTurn :: S.Set Int -> Face -> S.Set Int
+expandBlockTurn :: S.Set Int8 -> Face -> S.Set Int8
 expandBlockTurn block move
     | isBlockPreserved = rightSubSet
     | otherwise = allPieces
@@ -56,7 +57,7 @@ expandBlockTurn block move
         --There are going to be intersections. Must return the proper sub-partition
         rightSubSet = if (S.disjoint block xs2) then xs1 else xs2
 
-pieceIntersections :: (Foldable f) => f (S.Set Int) -> S.Set Int
+pieceIntersections :: (Foldable f) => f (S.Set Int8) -> S.Set Int8
 pieceIntersections = foldl' S.intersection (full)
     where
         full = allPieces
